Extract helper method
