+++ src/Workspaces/Core/Portable/FindSymbols/FindReferences/Finders/LinkedFileReferenceFinder.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Threading;
 using System.Threading.Tasks;
@@ -12,58 +11,11 @@ namespace Microsoft.CodeAnalysis.FindSymbols.Finders
 {
     internal class LinkedFileReferenceFinder : IReferenceFinder
     {
-        public async Task<ImmutableArray<SymbolAndProjectId>> DetermineCascadedSymbolsAsync(
+        public Task<ImmutableArray<SymbolAndProjectId>> DetermineCascadedSymbolsAsync(
             SymbolAndProjectId symbolAndProjectId, Solution solution, IImmutableSet<Project> projects,
             FindReferencesSearchOptions options, CancellationToken cancellationToken)
         {
-            var linkedSymbols = new HashSet<SymbolAndProjectId>();
-
-            var symbol = symbolAndProjectId.Symbol;
-            foreach (var location in symbol.DeclaringSyntaxReferences)
-            {
-                var originalDocument = solution.GetDocument(location.SyntaxTree);
-
-                // GetDocument will return null for locations in #load'ed trees.
-                // TODO:  Remove this check and add logic to fetch the #load'ed tree's
-                // Document once https://github.com/dotnet/roslyn/issues/5260 is fixed.
-                // TODO: the assert is also commented out becase generated syntax trees won't
-                // have a document until https://github.com/dotnet/roslyn/issues/42823 is fixed
-                if (originalDocument == null)
-                {
-                    // Debug.Assert(solution.Workspace.Kind == WorkspaceKind.Interactive || solution.Workspace.Kind == WorkspaceKind.MiscellaneousFiles);
-                    continue;
-                }
-
-                foreach (var linkedDocumentId in originalDocument.GetLinkedDocumentIds())
-                {
-                    var linkedDocument = solution.GetDocument(linkedDocumentId);
-                    var linkedSyntaxRoot = await linkedDocument.GetSyntaxRootAsync(cancellationToken).ConfigureAwait(false);
-
-                    // Defend against constructed solutions with inconsistent linked documents
-                    if (!linkedSyntaxRoot.FullSpan.Contains(location.Span))
-                    {
-                        continue;
-                    }
-
-                    var linkedNode = linkedSyntaxRoot.FindNode(location.Span, getInnermostNodeForTie: true);
-
-                    var semanticModel = await linkedDocument.GetSemanticModelAsync(cancellationToken).ConfigureAwait(false);
-                    var linkedSymbol = semanticModel.GetDeclaredSymbol(linkedNode, cancellationToken);
-
-                    if (linkedSymbol != null &&
-                        linkedSymbol.Kind == symbol.Kind &&
-                        linkedSymbol.Name == symbol.Name)
-                    {
-                        var linkedSymbolAndProjectId = SymbolAndProjectId.Create(linkedSymbol, linkedDocument.Project.Id);
-                        if (!linkedSymbols.Contains(linkedSymbolAndProjectId))
-                        {
-                            linkedSymbols.Add(linkedSymbolAndProjectId);
-                        }
-                    }
-                }
-            }
-
-            return linkedSymbols.ToImmutableArray();
+            return SymbolFinder.FindLinkedSymbolsAsync(symbolAndProjectId.Symbol, solution, cancellationToken);
         }
 
         public Task<ImmutableArray<Document>> DetermineDocumentsToSearchAsync(
